homepage

+++ backend/controllers/product.controller.js
@@ -35,7 +35,7 @@ export const updateProducts =async(req,res)=>{
     const productUpdate = req.body;
 
     if(!mongoose.Types.ObjectId.isValid(id)){
-        return res.status(404).jsono({status: false, message: "Invalid Product Id"});
+        return res.status(404).json({status: false, message: "Invalid Product Id"});
     }
 
     try {
@@ -50,11 +50,16 @@ export const updateProducts =async(req,res)=>{
 export const deleteProducts = async(req,res)=>{
     const {id}=req.params;
 
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({status: false, message: "Invalid Product Id"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message:"Product Deleted Successfully"})
     } catch (error) {
         console.error("Error in Deleting Product:", error.message);
-        res.status(404).json({status:false, message:"Product not found"})
+        res.status(500).json({status:false, message:"Server Error"})
     }
 }
